Added product service

+++ digimart_api/com.digimart/src/main/java/com/ecommerce/com/digimart/controllers/ProductController.java
@@ -1,11 +1,15 @@
 package com.ecommerce.com.digimart.controllers;
 
 import com.ecommerce.com.digimart.entities.Product;
+import com.ecommerce.com.digimart.services.impl.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/products")
@@ -20,7 +24,7 @@ public class ProductController
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Product> getProductById(@PathVariable Long id) {
+    public ResponseEntity<Product>  getProductById(@PathVariable int id) {
         Product product = productService.getProductById(id);
         return ResponseEntity.ok(product);
     }
